Jpa liquibase my sql

+++ src/main/java/se/implementer/muaythaiservice/model/FighterOverview.java
@@ -3,6 +3,7 @@ package se.implementer.muaythaiservice.model;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Value;
+import se.implementer.muaythaiservice.model.db.Fighter;
 import se.implementer.muaythaiservice.util.FighterUtil;
 
 import static se.implementer.muaythaiservice.model.FighterStatus.mapToFighterStatus;
@@ -13,7 +14,7 @@ import static se.implementer.muaythaiservice.model.Gender.mapToGender;
 @AllArgsConstructor
 public class FighterOverview {
 
-    int fighterId;
+    Long fighterId;
 
     String firstName;
